requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.2.0
pip==22.1.2
prettytable==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.3.0
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tornado==6.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
wcwidth==0.2.5
webdriver-manager==3.8.3
wheel==0.37.1
wsproto==1.1.0
